+++ taf/developer_tool.py
@@ -66,7 +66,6 @@ def add_roles(
     roles_key_infos=None,
     scheme=DEFAULT_RSA_SIGNATURE_SCHEME,
 ):
-
     yubikeys = defaultdict(dict)
     auth_repo = AuthenticationRepository(path=repo_path)
     repo_path = Path(repo_path)
@@ -157,7 +156,6 @@ def add_signing_key(
 
     parent_roles = set()
     for role in roles:
-
         if taf_repo.is_valid_metadata_key(role, pub_key_pem):
             print(f"Key already registered as signing key of role {role}")
             continue
@@ -329,13 +327,23 @@ def update_target_repos_from_repositories_json(
         "and specified in repositories.json"
     )
     for repo_name in repositories_json.get("repositories"):
-        target_repo_path = library_dir / repo_name
-        namespace_and_name = repo_name.rsplit("/", 1)
-        if len(namespace_and_name) > 1:
-            namespace, _ = namespace_and_name
-            targets_dir = auth_repo_targets_dir / namespace
-        targets_dir.mkdir(parents=True, exist_ok=True)
-        _update_target_repos(repo_path, targets_dir, target_repo_path, add_branch)
+        _save_top_commit_of_repo_to_target(
+            library_dir, repo_name, repo_path, add_branch
+        )
+
+
+def _save_top_commit_of_repo_to_target(
+    library_dir: Path, repo_name: str, auth_repo_path: Path, add_branch: bool = True
+):
+    auth_repo_targets_dir = auth_repo_path / TARGETS_DIRECTORY_NAME
+    target_repo_path = library_dir / repo_name
+    namespace_and_name = repo_name.rsplit("/", 1)
+    targets_dir = auth_repo_targets_dir
+    if len(namespace_and_name) > 1:
+        namespace, _ = namespace_and_name
+        targets_dir = auth_repo_targets_dir / namespace
+    targets_dir.mkdir(parents=True, exist_ok=True)
+    _update_target_repos(auth_repo_path, targets_dir, target_repo_path, add_branch)
 
 
 def _check_if_can_create_repository(auth_repo):
@@ -599,7 +607,6 @@ def _load_sorted_keys_of_new_roles(
 def _setup_roles_keys(
     role_name, key_info, repository, certs_dir=None, keystore=None, yubikeys=None
 ):
-
     yubikey_keys = []
     keystore_keys = []
 
@@ -908,7 +915,7 @@ def export_targets_history(repo_path, commit=None, output=None, target_repos=Non
         print(commits_json)
 
 
-def _get_namespace_and_root(repo_path, namespace, library_dir):
+def _get_namespace_and_root(repo_path, namespace=None, library_dir=None):
     repo_path = Path(repo_path).resolve()
     if namespace is None:
         namespace = repo_path.parent.name
@@ -1270,6 +1277,63 @@ def setup_test_yubikey(key_path=None):
     print("\nPrivate key successfully imported.\n")
     print("\nPublic key (PEM): \n{}".format(pub_key.decode("utf-8")))
     print("Pin: {}\n".format(pin))
+
+
+def update_and_sign_targets(
+    repo_path: str,
+    library_dir: str,
+    target_types: list,
+    keystore: str,
+    roles_key_infos: str,
+    scheme: str,
+):
+    """
+    <Purpose>
+        Save the top commit of specified target repositories to the corresponding target files and sign
+    <Arguments>
+        repo_path:
+        Authentication repository's location
+        library_dir:
+        Directory where target repositories and, optionally, authentication repository are locate
+        targets:
+        Types of target repositories whose corresponding target files should be updated and signed
+        keystore:
+        Location of the keystore files
+        roles_key_infos:
+        A dictionary whose keys are role names, while values contain information about the keys
+        no_commit:
+        Indicates that the changes should bot get committed automatically
+        scheme:
+        A signature scheme used for signing
+
+    """
+    auth_path = Path(repo_path).resolve()
+    auth_repo = AuthenticationRepository(path=auth_path)
+    if library_dir is None:
+        library_dir = auth_path.parent.parent
+    repositoriesdb.load_repositories(auth_repo)
+    nonexistent_target_types = []
+    target_names = []
+    for target_type in target_types:
+        try:
+            target_name = repositoriesdb.get_repositories_paths_by_custom_data(
+                auth_repo, type=target_type
+            )[0]
+            target_names.append(target_name)
+        except Exception:
+            nonexistent_target_types.append(target_type)
+            continue
+    if len(nonexistent_target_types):
+        print(
+            f"Target types {'.'.join(nonexistent_target_types)} not in repositories.json. Targets not updated"
+        )
+        return
+
+    # only update target files if all specified types are valid
+    for target_name in target_names:
+        _save_top_commit_of_repo_to_target(library_dir, target_name, auth_path, True)
+        print(f"Updated {target_name} target file")
+    register_target_files(auth_path, keystore, roles_key_infos, True, scheme)
 
 
 def update_metadata_expiration_date(